# api/openapi/mint-server.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "mint_server/agent/agent.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthService_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/authLoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/authLoginRequest"
            }
          }
        ],
        "tags": [
          "AuthService"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthService_RegisterAccount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/authRegisterAccountResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/authRegisterAccountRequest"
            }
          }
        ],
        "tags": [
          "AuthService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "authLoginRequest": {
      "type": "object",
      "properties": {
        "provider": {
          "type": "string",
          "title": "登录方式，当前支持 account_password"
        },
        "identifier": {
          "type": "string",
          "title": "登录标识，账号登录时为账号名"
        },
        "credential": {
          "type": "string",
          "title": "登录凭证，账号密码登录时为明文密码"
        }
      }
    },
    "authLoginResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "title": "业务状态码，0 表示成功"
        },
        "message": {
          "type": "string",
          "title": "业务响应消息"
        },
        "accessToken": {
          "type": "string",
          "title": "JWT 访问令牌"
        },
        "tokenType": {
          "type": "string",
          "title": "令牌类型，固定为 Bearer"
        },
        "expiresIn": {
          "type": "string",
          "format": "int64",
          "title": "令牌过期秒数"
        },
        "user": {
          "$ref": "#/definitions/mint_serverAuthUserInfo",
          "title": "登录用户信息"
        }
      }
    },
    "authRegisterAccountRequest": {
      "type": "object",
      "properties": {
        "account": {
          "type": "string",
          "title": "账号密码注册使用的唯一账号"
        },
        "password": {
          "type": "string",
          "title": "账号密码注册使用的明文密码"
        },
        "displayName": {
          "type": "string",
          "title": "用户展示名称"
        },
        "avatarUrl": {
          "type": "string",
          "title": "用户头像地址"
        }
      }
    },
    "authRegisterAccountResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "title": "业务状态码，0 表示成功"
        },
        "message": {
          "type": "string",
          "title": "业务响应消息"
        },
        "accessToken": {
          "type": "string",
          "title": "JWT 访问令牌"
        },
        "tokenType": {
          "type": "string",
          "title": "令牌类型，固定为 Bearer"
        },
        "expiresIn": {
          "type": "string",
          "format": "int64",
          "title": "令牌过期秒数"
        },
        "user": {
          "$ref": "#/definitions/mint_serverAuthUserInfo",
          "title": "注册用户信息"
        }
      }
    },
...
    "mint_serverAuthUserInfo": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "title": "用户ID"
        },
        "displayName": {
          "type": "string",
          "title": "用户展示名称"
        },
        "avatarUrl": {
          "type": "string",
          "title": "用户头像地址"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "title": "用户状态，1 表示启用"
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
  }
}